Trying to librify the meiosis integration with superouter
